+++ CHANGELOG.md
@@ -8,6 +8,13 @@ this changelog also includes the original WARDEN changelog.
 * Fix infinite recursion on clock conversion
 * Integration tests with default validity periods
 * Fix wrong offset sign with secondary constructor
+* Rework NOOP attestation and NOOP results
+    * Non-error AttestationResults now come with an `AttestationResult.Kind` marker interface
+    * Makoto produces `AttestationResult.Verified`
+    * NoopAttestationService produces `AttestationResult.NOOP`
+    * `KeyAttestation.fold` now produces a nullable `AttestationResult.Verified` on success to acccount for NOOP results
+    * Makoto and NoopAttestationService bring their own `foldTyped` extenstion (which sadly cannot override a common abstract extension, because they need to be inline)
+* make `makoto` property of AttestationVerifier public
 * Versioned debug statements
 
 ## 0.9.999